etls/requirements.txt
mysql-connector-python==8.0.21
requests==2.24.0
pika==1.1.0
datetime